requirements.txt
langchain>=0.2.0
langchain-community>=0.0.20
openai
unstructured[all-docs]
pypdf
python-docx
python-pptx
pandas
openpyxl
xlrd
striprtf
pillow
pytesseract
google-auth
google-api-python-client
google-auth-httplib2
google-auth-oauthlib
chromadb
faiss-cpu
tiktoken
PyMuPDF
tqdm
tabulate
gradio
pdfminer.six
pi-heif
unstructured_inference
tk
